+++ src/commands/hardis/work/new.ts
@@ -59,13 +59,21 @@ export default class NewTask extends SfdxCommand {
     await checkGitClean({ allowStash: true });
 
     const config = await getConfig("project");
+    const availableTargetBranches = config.availableTargetBranches || null;
 
     // Request info to build branch name. ex features/config/MYTASK
     const response = await prompts([
       {
-        type: "text",
+        type: availableTargetBranches ? "select" : "text",
         name: "targetBranch",
-        message: c.cyanBright("What will be the target branch of your new task ? (the branch where you will make your merge request after the task is completed)"),
+        message: c.cyanBright(
+          "What will be the target branch of your new task ? (the branch where you will make your merge request after the task is completed)"
+        ),
+        choices: availableTargetBranches
+          ? availableTargetBranches.map((branch) => {
+              return { title: branch, value: branch };
+            })
+          : [],
         initial: config.developmentBranch || "developpement",
       },
       {
@@ -113,13 +121,22 @@ export default class NewTask extends SfdxCommand {
 
     // Update config if necessary
     if (config.developmentBranch !== targetBranch) {
-      await setConfig("project", { developmentBranch: targetBranch });
+      const updateDefaultBranchRes = await prompts({
+        type: "confirm",
+        name: "value",
+        message: c.cyanBright(`Do you want to update the default development branch to ${c.green(targetBranch)} ?`),
+        default: false,
+      });
+      if (updateDefaultBranchRes.value === true) {
+        await setConfig("project", { developmentBranch: targetBranch });
+      }
     }
 
     // Select/Create scratch org
     const currentOrg = await MetadataUtils.getCurrentOrg("scratch");
     if (currentOrg == null) {
-      const scratchOrgList = await MetadataUtils.listLocalOrgs("scratch");
+      const hubOrgUsername = this?.hubOrg?.getUsername();
+      const scratchOrgList = await MetadataUtils.listLocalOrgs("scratch", { devHubUsername: hubOrgUsername });
       const scratchResponse = await prompts({
         type: "select",
         name: "value",